Encapsulate &&&PhantomData syntax behind wrapper macros
And add a note in the docs that you need to use those macros instead of calling the user-facing traits
