+++ Cargo.toml
@@ -14,7 +14,7 @@ futures = { version = "0.3.13", default-features = false } # for async in genera
 chrono = { version = "0.4.19", default-features = false } # interoperate with serenity timestamps
 once_cell = { version = "1.7.2", default-features = false, features = ["std"] } # to store and set user data
 poise_macros = { path = "macros" }
-async-trait = { version = "0.1.48", default-features = false } # PopArgumentAsync trait
+async-trait = { version = "0.1.48", default-features = false } # various traits
 regex = { version = "1.5.4", default-features = false, features = ["std"] } # prefix
 
 [dependencies.serenity]
