+++ src/slash/argument/mod.rs
@@ -1,7 +1,10 @@
 //! Application command argument handling code
 
-mod slash;
-pub use slash::*;
+mod slash_macro;
+pub use slash_macro::*;
+
+mod slash_trait;
+pub use slash_trait::*;
 
 mod context_menu;
 pub use context_menu::*;
